requirements_orig.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Login>=0.6
Flask-Mail>=0.10
itsdangerous>=2.1
Werkzeug>=3.0
python-dotenv>=1.0
requests>=2.31
APScheduler>=3.10
